requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cramjam==2.7.0
fastparquet==2023.10.1
fsspec==2023.12.0
google-api-core==2.14.0
google-auth==2.25.1
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
idna==3.6
numpy==1.26.2
packaging==23.2
pandas==2.1.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
